Finish New FOW
Missing tags/layers/colliders needed in rooms prefab for the new FOW to work. Tested and was fully fonctional. The only issue while testing was the FOW going throught doors ( i think it was doors ) meaning there was either no box collider ( should not be the case ) or missing tag/layer. To merge the new FOW, use the new FogOfWar prefab, position : 58.7x, 59.2y, 0 or 10z. Parameters might need to be set. Modifications needed in the playercontrol script for the FOW to work and add the missing tag/layer in the rooms prefabs.